+++ src/ast/passes/semantic_analyser.cpp
@@ -595,6 +595,12 @@ static const std::map<std::string, call_spec> CALL_SPEC = {
       .max_args=1,
       .discard_ret_warn = true },
   },
+  { "socket_cookie",
+    { .min_args=1,
+      .max_args=1,
+      .arg_types={
+        arg_type_spec{ .type=Type::pointer }, // struct sock *
+      } } },
 };
 // clang-format on
 
@@ -1753,6 +1759,24 @@ If you're seeing errors, try clamping the string sizes. For example:
                         << arg.type().GetTy() << " provided)";
       }
     }
+  } else if (call.func == "socket_cookie") {
+    auto logError = [&]<typename T>(T name) {
+      call.addError() << call.func
+                      << "() only supports 'struct sock *' as the argument ("
+                      << name << " provided)";
+    };
+
+    const auto &type = call.vargs.at(0).type();
+    if (!type.IsPtrTy() || !type.GetPointeeTy() ||
+        !type.GetPointeeTy()->IsRecordTy()) {
+      logError(type.GetTy());
+      return;
+    }
+    if (!type.GetPointeeTy()->IsSameType(CreateRecord("struct sock"))) {
+      logError("'" + type.GetPointeeTy()->GetName() + " *'");
+      return;
+    }
+    call.return_type = CreateUInt64();
   } else {
     call.addError() << "Unknown function: '" << call.func << "'";
   }