+++ CHANGELOG.md
@@ -65,6 +65,8 @@ and this project adheres to
   - [#4317](https://github.com/bpftrace/bpftrace/pull/4317)
 - Add `getopt` function to handle named command line program arguments
   - [#4258](https://github.com/bpftrace/bpftrace/pull/4258)
+- Add `socket_cookie` function to get the cookie of a socket
+  - [#4345](https://github.com/bpftrace/bpftrace/pull/4345)
 #### Changed
 - `-p` CLI flag now applies to all probes (except BEGIN/END)
   - [#3800](https://github.com/bpftrace/bpftrace/pull/3800)